requirements.txt
dj-database-url
Django
whitenoise
gunicorn
psycopg2
requests
djangorestframework
